(spotbugs) fix some problems, replace class w/ record
